v2.0.5-rc (#41)
Reviewed-on: https://git.fifthgrid.com/BlockStorage/repertory/pulls/41

+++ scripts/docker_common.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
-if [ "${PROJECT_BUILD_ARCH}" == "aarch64" ]; then
+if [ "${PROJECT_BUILD_ARCH}" == "aarch64" ] &&
+  [ "${PROJECT_ENABLE_MULTIARCH_DOCKER}" == "1" ]; then
   docker run --rm --privileged multiarch/qemu-user-static --reset -p yes
 fi
